+++ application/todo-list/app/models/user.rb
@@ -2,15 +2,13 @@ class User < ActiveRecord::Base
   validates :username, presence: true, uniqueness: true
   validates :password, presence: true
 
-  @authenticated = false
+  has_many :lists
 
-  def authenticate!(password)
-    if self.password == password
-      @authenticated = true
-    end
+  def will_authenticate?(password)
+    self.password == password
   end
 
-  def is_authenticated?()
-    @authenticated
+  def logout!()
+    puts "INFO: #{username} logs out."
   end
 end