+++ application/todo-list/db/migrate/20140404201133_create_lists.rb
@@ -1,0 +1,12 @@
+class CreateLists < ActiveRecord::Migration
+  def change
+    create_table :lists do |t|
+      t.belongs_to :user
+      t.string :name
+      t.date :date
+      t.boolean :available, :default => true
+
+      t.timestamps
+    end
+  end
+end