+++ application/todo-list/db/schema.rb
@@ -11,7 +11,25 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20140404194812) do
+ActiveRecord::Schema.define(version: 20140404201351) do
+
+  create_table "lists", force: true do |t|
+    t.integer  "user_id"
+    t.string   "name"
+    t.date     "date"
+    t.boolean  "available",  default: true
+    t.datetime "created_at"
+    t.datetime "updated_at"
+  end
+
+  create_table "tasks", force: true do |t|
+    t.integer  "list_id"
+    t.string   "title"
+    t.date     "date"
+    t.boolean  "closed",     default: false
+    t.datetime "created_at"
+    t.datetime "updated_at"
+  end
 
   create_table "users", force: true do |t|
     t.string   "username"