+++ application/todo-list/features/step_definitions/tasks.rb
@@ -1,27 +1,26 @@
 Then(/^"([^"]+)" item will be closed$/) do |taskName|
-  pending
+  task = createTaskIfNotExist(taskName)
+  task.close!
+  task.save!
 end
 
 Then(/^"([^"]+)" item will be opened$/) do |taskName|
-  pending
+  task = createTaskIfNotExist(taskName)
+  task.save!
 end
 
 Then(/^"([^"]+)" item should be opened$/) do |taskName|
-  pending
+  Task.where(:title => taskName).first.closed.should == false
 end
 
 Then(/^"([^"]+)" item should be closed$/) do |taskName|
-  pending
-end
-
-Then(/^"([^"]+)" item should be available$/) do |taskName|
-  pending
+  Task.where(:title => taskName).first.closed.should == true
 end
 
 Then(/^(\d+) task in "([^"]+)" is available$/) do |tasksNumber, listName|
-  pending
+  List.where(:name => listName).first.tasks.count.should == tasksNumber.to_i
 end
 
 Then(/^"([^"]+)" item should have today's date$/) do |taskName|
-  pending
+  Task.where(:title => taskName).first.date.should == Date.today
 end