+++ application/todo-list/features/step_definitions/users.rb
@@ -1,5 +1,5 @@
-Given(/^"([^"]+)" should have "([^"]+)"$/) do |username, list|
-  pending
+Given(/^"([^"]+)" should have "([^"]+)"$/) do |username, listName|
+  List.where(:name => listName).first.user.username.should == username
 end
 
 Given(/^there are no users$/) do
@@ -11,39 +11,49 @@ Given(/^there is one user "([^"]+)" with password "([^"]+)"$/) do |username, pas
 end
 
 Given(/^there is one authenticated user "([^"]+)" with password "([^"]+)"$/) do |username, password|
-  pending
+  user = User.create(:username => username, :password => password)
+  user.save!
+
+  if !user.will_authenticate?(password)
+    fail("Invalid password for '#{username}'!")
+  end
 end
 
 When(/^"([^"]+)" account is created with random password$/) do |username|
   User.create!(:username => username, :password => random_password())
 end
 
-When(/^authenticate "([^"]+)" with password "([^"]+)"$/) do |username, password|
-  User.where(:username => username).first.authenticate!(password)
+When(/^"([^"]+)" log out$/) do |username|
+  User.where(:username => username).first.logout!
 end
 
-When(/^"([^"]+)" log out$/) do |username|
-  pending
+Then(/^authenticate "(.*?)" with password "(.*?)"$/) do |username, password|
+  user = User.where(:username => username).first
+
+  if !user.will_authenticate?(password)
+    fail("Invalid password for '#{username}'!")
+  end
 end
 
 Then(/^"([^"]+)" should have non empty password$/) do |username|
   user = User.where(:username => username).first
 
-  user.password && !user.password.empty?
+  user.password.should_not == nil
+  user.password.empty?.should == false
 end
 
-Then(/^"([^"]+)" should be authenticated$/) do |username|
-  User.where(:username => username).first.is_authenticated?
+Then(/^"([^"]+)" should be authenticated with password "([^"]+)"$/) do |username, password|
+  User.where(:username => username).first.will_authenticate?(password).should == true
 end
 
-Then(/^"([^"]+)" should not be authenticated$/) do |username|
-  !User.where(:username => username).first.is_authenticated?
+Then(/^"([^"]+)" should not be authenticated with password "([^"]+)"$/) do |username, password|
+  User.where(:username => username).first.will_authenticate?(password).should == false
 end
 
 Then(/^"([^"]+)" account should be available$/) do |username|
-  User.where(:username => username).count == 1
+  User.where(:username => username).count.should == 1
 end
 
 Then(/^"([^"]+)" should have no lists$/) do |username|
-  pending
+  User.joins(:lists).where("users.username == ?", username).count.should == 0
 end