+++ application/todo-list/spec/models/list_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+
+describe List do
+  it { should have_many :tasks }
+end
+
+describe List do
+  it { should belong_to :user }
+end