+++ application/Changelog.md
@@ -2,3 +2,13 @@
 2. `gem install bundler`
 3. `bundle install`
 4. `rails ./todo-list`
+5. `rails generate model User username:string password:string`
+6. `rake db:migrate`
+7. `rake db:test:clone`
+8. `cucumber features/User\ Management.feature:6`
+9. `rails generate model List name:string datetime:date available:boolean`
+10. `rails generate model Task title:string datetime:date state:boolean`
+11. `rake db:migrate`
+12. `rake db:test:clone`
+13. `rails generate rspec:install`
+14. `rake spec`